+++ examples/src/NonDet.hs
@@ -1,15 +1,15 @@
 {-# LANGUAGE FlexibleContexts #-}
-module NonDetEff where
+module NonDet where
 
 import Control.Applicative
 import Control.Monad
 import Control.Monad.Freer
 
-ifte :: Member NonDetEff r
+ifte :: Member NonDet r
      => Eff r a -> (a -> Eff r b) -> Eff r b -> Eff r b
 ifte t th el = (t >>= th) <|> el
 
-testIfte :: Member NonDetEff r => Eff r Int
+testIfte :: Member NonDet r => Eff r Int
 testIfte = do
   n <- gen
   ifte (do d <- gen